Move rest of implementations of CallInterceptor to internal-instrumentation-api project
