+++ subprojects/core/src/main/java/org/gradle/internal/classpath/intercept/CallSiteInterceptorSet.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023 the original author or authors.
+ * Copyright 2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
